checkio problem to remove all unique items from a list
Remove unique items from a list and maintain ordering. I used an integer that starts with a positive value. This only decreases when an iteration over the list produces no instances in finding unique numbers. This is used to control a boolean "switch" to break out of the while loop encapsulating the for loop.